test(sessions): trim maintenance cap duplicates
